Added base work for messaging
Added the message file which contains functions related to messaging, then added to the app.
