+++ app/app_settings.json
@@ -1,3 +1,3 @@
 {
-  "appName": "insertwork"
+  "appName": "inserttool"
 }
